+++ src/app/views/certificates/ResponseDialog.js
@@ -8,14 +8,15 @@ import DialogTitle from '@material-ui/core/DialogTitle';
 import MUIDataTable from "mui-datatables";
 import { createMuiTheme, MuiThemeProvider } from '@material-ui/core/styles';
 
-import Table from '@material-ui/core/Table';
-import TableBody from '@material-ui/core/TableBody';
-import TableCell from '@material-ui/core/TableCell';
-import TableContainer from '@material-ui/core/TableContainer';
-import TableHead from '@material-ui/core/TableHead';
-import TableRow from '@material-ui/core/TableRow';
-import Paper from '@material-ui/core/Paper';
-import { makeStyles } from '@material-ui/core/styles';
+// import Table from '@material-ui/core/Table';
+// import TableBody from '@material-ui/core/TableBody';
+// import TableCell from '@material-ui/core/TableCell';
+// import TableContainer from '@material-ui/core/TableContainer';
+// import TableHead from '@material-ui/core/TableHead';
+// import TableRow from '@material-ui/core/TableRow';
+// import Paper from '@material-ui/core/Paper';
+// import { makeStyles } from '@material-ui/core/styles';
+import CertificatesResult from './certificates-utils/CertificatesResults'
 
 export default function ResponseDialog({ openDialog, setOpenDialog, responseData, isLoading }) {
 
@@ -36,6 +37,8 @@ export default function ResponseDialog({ openDialog, setOpenDialog, responseData
   return (
     <div>
       <Dialog
+        maxWidth={'xl'}
+        fullWidth={true}
         open={openDialog}
         onClose={handleClose}
         scroll={"paper"}
@@ -43,7 +46,7 @@ export default function ResponseDialog({ openDialog, setOpenDialog, responseData
         aria-describedby="scroll-dialog-description"
       >
         <DialogTitle id="scroll-dialog-title">Results</DialogTitle>
-        <DialogContent dividers={true}>
+        <DialogContent dividers={true} >
           <DialogContentText
             children={<ResponseContent responseData={responseData} isLoading={isLoading} />}
             id="scroll-dialog-description"
@@ -64,35 +67,6 @@ export default function ResponseDialog({ openDialog, setOpenDialog, responseData
 
 const ResponseContent = ({ responseData, isLoading }) => {
 
-    const useStyles = makeStyles({
-  table: {
-    minWidth: 650,
-  },
-});
-
-    const classes = useStyles();
-
-
-  const getMuiThemeError = () => createMuiTheme({
-    overrides: {
-      MUIDataTableBodyCell: {
-        root: {
-          backgroundColor: "#FF0000"
-        }
-      }
-    }
-  })
-
-  const getMuiThemeSuccess = () => createMuiTheme({
-    overrides: {
-      MUIDataTableBodyCell: {
-        root: {
-          backgroundColor: "#32CD32"
-        }
-      }
-    }
-  })
-
   const { data } = responseData
 
   const successData = data.success.map((item) => {
@@ -109,84 +83,9 @@ const ResponseContent = ({ responseData, isLoading }) => {
       message: item.msg
     }
   })
-
-  console.log("errorData:", errorData)
-
-  const columnsSuccess = [
-    {
-      name: "student",
-      label: "student",
-
-    },
-    {
-      name: "status",
-      label: "status",
-
-    },
-    {
-      name: "message",
-      label: "message",
-
-    },
-  ]
-  const columnsError = [
-    {
-      name: "student",
-      label: "student",
-
-    },
-    {
-      name: "status",
-      label: "status",
-
-    },
-    {
-      name: "message",
-      label: "message",
-
-    },
-  ]
-
+  
   return (
-    <>
-    <TableContainer component={Paper}>
-      <Table className={classes.table} aria-label="simple table">
-        <TableHead>
-          <TableRow>
-            <TableCell align="right">Student</TableCell>
-            <TableCell align="right">Status</TableCell>
-            <TableCell align="right">Message</TableCell>
-          </TableRow>
-        </TableHead>
-        <TableBody>
-          {errorData.map((row) => (
-            <TableRow key={row.name}>
-              <TableCell component="th" scope="row">
-                {row.name}
-              </TableCell>
-              <TableCell >{row.student}</TableCell>
-              <TableCell >{row.status}</TableCell>
-              <TableCell >{row.message}</TableCell>
-            </TableRow>
-          ))}
-        </TableBody>
-      </Table>
-    </TableContainer>
-      {/* <MuiThemeProvider theme={getMuiThemeSuccess()}>
-        <MUIDataTable
-          title={"Certificates Created"}
-          data={successData}
-          columns={columnsSuccess}
-        />
-      </MuiThemeProvider> */}
-      {/* <MuiThemeProvider theme={getMuiThemeError()}>
-        <MUIDataTable
-          title={"Pending Certificates"}
-          data={errorData}
-          columns={columnsError}
-        />
-      </MuiThemeProvider> */}
-    </>
+    <CertificatesResult errorData={errorData} successData={successData} />
   )
 }
