+++ src/app/views/certificates/certificates-utils/CertificatesResults.js
@@ -1,0 +1,56 @@
+import React from 'react';
+import { makeStyles } from '@material-ui/core/styles';
+import Table from '@material-ui/core/Table';
+import TableBody from '@material-ui/core/TableBody';
+import TableCell from '@material-ui/core/TableCell';
+import TableContainer from '@material-ui/core/TableContainer';
+import TableHead from '@material-ui/core/TableHead';
+import TableRow from '@material-ui/core/TableRow';
+import Paper from '@material-ui/core/Paper';
+
+const useStyles = makeStyles({
+  table: {
+    minWidth: 650,
+  },
+});
+
+function createData(name, calories, fat) {
+  return { name, calories, fat};
+}
+
+const rows = [
+  createData('Frozen yoghurt', 159, 6.0),
+  createData('Ice cream sandwich', 237, 9.0),
+  createData('Eclair', 262, 16.0),
+];
+
+export default function CertificatesResult({errorData, successData}) {
+  const classes = useStyles();
+
+  const rows = [...errorData, ...successData]
+
+  return (
+    <TableContainer component={Paper}>
+      <Table className={classes.table} aria-label="simple table">
+        <TableHead>
+          <TableRow>
+            <TableCell>Student</TableCell>
+            <TableCell >Status</TableCell>
+            <TableCell >Message</TableCell>
+          </TableRow>
+        </TableHead>
+        <TableBody>
+          {rows.map((row) => (
+            <TableRow key={row.name}>
+              <TableCell component="th" scope="row">
+                {row.student}
+              </TableCell>
+              <TableCell >{row.status}</TableCell>
+              <TableCell >{row.message}</TableCell>
+            </TableRow>
+          ))}
+        </TableBody>
+      </Table>
+    </TableContainer>
+  );
+}
